Added readme.md
# OpenAI Console Application

This is a console application that uses OpenAI's GPT/DaVinci model to generate text based on text-based input.

## Usage

1. Replace the API key and path to OpenAI's API in the program file.
2. Run the console application.
3. The generated text will be displayed in the console.

## Dependencies

This application uses the following NuGet packages:

- Serilog
- Serilog.Sinks.Console
- IronPython
- Microsoft.Scripting

Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.

## License

This application is licensed under the MIT License.